requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.1
google-genai>=0.3.0
tenacity>=8.2.3
